apps/mcp-server/requirements.txt
fastmcp>=2.0
httpx
supabase
pydantic
uvicorn
python-dotenv
